pyproject.toml
[tool.poetry]
name = "rev_ocpp"
version = "0.1.0"
description = "Charging station management software"
authors = ["Travis <tgpovey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
SQLAlchemy = "^1.3.23"
ocpp = "^0.8.1"
alembic = "^1.5.6"
psycopg2 = "^2.8.6"
pymongo = "^3.11.3"
uvicorn = "^0.13.4"
gunicorn = "^20.0.4"
python-multipart = "^0.0.5"
pydantic = {extras = ["email"], version = "^1.8.1"}
python-jose = {extras = ["cryptography"], version = "^3.2.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
